refactoring: move the strip_user_mention_display function to comment completion.
